Added clicker minigame
Refactoring of battle/fight page layout to include health, time left, amount of clicks textviews and fight/hit buttons needed for the clicker minigame. The amount of clicks in a period of 5 seconds  will now be registered and used to calculate a random amount of damage and will be subtracted from the health textviews.