Added highscore functionality
When ending the clicker minigame it will now check if your score has made it in the top 15. If it did it will remove the lowest score in the database and add your score. Also added a page to view all the highscore wich uses a recyclerview to display everything.